+++ Sources/ZcashLightClientKit/DAO/PagedTransactionDao.swift
@@ -13,17 +13,21 @@ class PagedTransactionDAO: PaginatedTransactionRepository {
     var kind: TransactionKind
     
     var pageCount: Int {
-        guard pageSize > 0 else {
-            return 0
+        get async {
+            guard pageSize > 0 else {
+                return 0
+            }
+            return await itemCount / pageSize
         }
-        return itemCount / pageSize
     }
     
     var itemCount: Int {
-        guard let count = try? transactionRepository.countAll() else {
-            return 0
+        get async {
+            guard let count = try? await transactionRepository.countAll() else {
+                return 0
+            }
+            return count
         }
-        return count
     }
     
     init(repository: TransactionRepository, pageSize: Int = 30, kind: TransactionKind = .all) {
@@ -32,17 +36,16 @@ class PagedTransactionDAO: PaginatedTransactionRepository {
         self.kind = kind
     }
     
-    func page(_ number: Int) throws -> [ZcashTransaction.Overview]? {
+    func page(_ number: Int) async throws -> [ZcashTransaction.Overview]? {
         let offset = number * pageSize
-        guard offset < itemCount else { return nil }
-        return try transactionRepository.find(offset: offset, limit: pageSize, kind: kind)
+        guard offset < (await itemCount) else { return nil }
+        return try await transactionRepository.find(offset: offset, limit: pageSize, kind: kind)
     }
     
     func page(_ number: Int, result: @escaping (Result<[ZcashTransaction.Overview]?, Error>) -> Void) {
-        DispatchQueue.global(qos: .userInitiated).async { [weak self] in
-            guard let self else { return }
+        Task(priority: .userInitiated) {
             do {
-                result(.success(try self.page(number)))
+                result(.success(try await self.page(number)))
             } catch {
                 result(.failure(error))
             }